consolidation conditions
