+++ tuxemon/event/conditions/player_at.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2014-2023 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
 from __future__ import annotations
 
-from tuxemon.event import MapCondition
+from tuxemon.event import MapCondition, collide
 from tuxemon.event.eventcondition import EventCondition
 from tuxemon.session import Session
 
@@ -33,15 +33,4 @@ class PlayerAtCondition(EventCondition):
 
         """
         player = session.player
-
-        # Get the condition's rectangle area. If we're on a tile in that area,
-        # then this condition should return True.
-        area_x = range(condition.x, condition.x + condition.width)
-        area_y = range(condition.y, condition.y + condition.height)
-
-        # If the player is at the coordinates and the operator is set to true
-        # then return true
-        return (
-            round(player.tile_pos[0]) in area_x
-            and round(player.tile_pos[1]) in area_y
-        )
+        return collide(condition, player.tile_pos)
